Don't take square root when calculating whether to escape calculation
If the magnitude is less than or equal to 2, bail on the calculation.
We have determined that the number is outside the set. However, this
reqires doing a square root. Instead, just use the squared magnitude
and compare that to 4. An identical determination is made but wih much
less work being performed. A local benchmark (on battery on this old
2012 laptop) showed an FPS bunch from 12fps to 15fps with recalculation
always happening. Glad to see such a big perf win.